Move vscode-dependent utility functions to their own file to prevent test runner issues
